pyproject.toml
[project]
name = "image-manager"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = [
    "alembic>=1.16.1",
    "fastapi[standard]>=0.115.12",
    "google-cloud-storage>=3.1.0",
    "psycopg2>=2.9.10",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.34.2",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "ruff>=0.11.11",
]
